refactor: structure library internally
