+++ src/parse/mod.rs
@@ -1,0 +1,4 @@
+pub mod parser;
+
+pub(crate) use parser::{CommandKind, ImpKind, ParamKind};
+pub use parser::{Instruction, Parser};
